performance-tuning-roadmap/labs/perfshop-p0/app/requirements.txt
mysql-connector-python==9.0.0

